+++ app/src/test/java/moe/antimony/hoshi/LocalizationResourceTest.kt
@@ -115,6 +115,7 @@ class LocalizationResourceTest {
         assertEquals("标注", zhResources.strings.getValue("reader_highlight_action").value)
         assertEquals("复制", zhResources.strings.getValue("action_copy").value)
         assertEquals("分享", zhResources.strings.getValue("action_share").value)
+        assertEquals("段落间距", zhResources.strings.getValue("reader_appearance_paragraph_spacing").value)
     }
 
     @Test